#2 Filtering options user name search
Add `search_users/3` function to use the levenshtein algorithm. We default our max_distance to 3 as an optional field. This allows a search of "Jer Brayton" to find me since 3 letters were removed. Merchant search should be identical with only one field instead of 2.

It's worth noting that the search test doesn't cover all that much but I want to circle back around to those later. If we're going to test mutations and queries, we could cover it all then.