#2 Filtering options transaction amount between min and max
Implement `search_transactions/2` that looks for transaction amounts between min and max. The GraphQL query arguments have min as an optional field, defaulting to zero. It could be useful to default the repo function but min after max is a weird aesthetic. The function defaulting to `func(max)` makes sense but `func(max, min)` is very easy for me to transpose.

Our tests are a little better as we handle 2 different transactions. We could've also ran a test to grab either the 42 or 43 amount.