#3 Transactions in decimal default_value
Took an exceedingly long time to research the problem as it wasn't all that obvious. `Decimal.to_string/2` is called as `Decimal.to_string("0.00", :scientific)` but it apparently did not match the other function overloads starting at decimal.ex:1379.

The error in GraphiQL:
```
** (FunctionClauseError) no function clause matching in Decimal.to_string/2

    The following arguments were given to Decimal.to_string/2:

        # 1
        "0.00"

        # 2
        :scientific
```

Stacktrace 1st line in the log:
```
(decimal 1.8.1) lib/decimal.ex:1379: Decimal.to_string("0.00", :scientific)
```

Finally realized I could pass in either the `%Decimal{}` struct or in this instance we're initializing a decimal type as zero.
Refactored the `search_transactions` function in the resolver to reduce the temporary variables. I believe I needed that at a single point when the bug popped up.