#4 Company entities relationships
Migrations to add our company_id foreign key to the `users` and `transactions` tables.
Relationships added to the Repo schemas and tests altered to pass in company identifiers.
User tests needed to be altered to conform to the transactions test `setup` section to get around sandbox issues.
Add field resolvers and `company_id` to both Users and Transactions GraphQL layer.

**Note: While the docs don't explicitly call out a relationship for transactions, I made an executive decision to include it as a foreign key. The UUID detached for historical purposes makes sense if this could be used when users change companies. This is rife for discussion as the approach hinges on the behavior of the overall system. I've rarely seen a system where users can't change companies so I reserve the right to revise this decision later.**

The error returned when trying to use `Companies.create_company` outside the setup block:
```
== Compilation error in file test/homework/users_test.exs ==
** (DBConnection.OwnershipError) cannot find ownership process for #PID<0.453.0>.

When using ownership, you must manage connections in one
of the four ways:

* By explicitly checking out a connection
* By explicitly allowing a spawned process
* By running the pool in shared mode
* By using :caller option with allowed process

The first two options require every new process to explicitly
check a connection out or be allowed by calling checkout or
allow respectively.

The third option requires a {:shared, pid} mode to be set.
If using shared mode in tests, make sure your tests are not
async.

The fourth option requires [caller: pid] to be used when
checking out a connection from the pool. The caller process
should already be allowed on a connection.

If you are reading this error, it means you have not done one
of the steps above or that the owner process has crashed.

See Ecto.Adapters.SQL.Sandbox docs for more information.
    (ecto_sql 3.4.5) lib/ecto/adapters/sql.ex:590: Ecto.Adapters.SQL.raise_sql_call_error/1
    (ecto 3.4.6) lib/ecto/repo/schema.ex:661: Ecto.Repo.Schema.apply/4
```