#7 Bonus transaction bug
Seems the credit field was missing from the cast. We remove both that and the debit from validation checking as they both default to false.
To treat the transactions as a ledger, it would be ideal to set either credit or debit to true as I believe only one should be true while the other is false at any given time. I can't think of a case where you'd want both debit and credit as true or false. I do hope to get to that later but I'm looking to get more tests created first.