#5 Test resolvers and mutations for merchants
Enable `/api` endpoint as well as Absinthe parser to work with tools like Insomnia. Tried specifying the parser in the router following the docs before realizing the existing setup was in the endpoint. That took longer to get working than I would've liked but I was determined to use tooling I'm familiar with.
Merchant tests cover the happy paths of the resolver. It would be worth investing in covering the invalid paths but I'll look to do that last if time permits.