#5 Test resolvers and mutations for companies, converting to decimal
Convert company amounts to decimals with a tiny smell. It would be useful to extract the base to_dollars and to_cents functions that work with the Decimal usage to another module but I typically use the rule of three for refactoring. It's an area for improvement for sure but I spent a bit of time before I realized I needed to change the schema.

This also exposes the problem with having this at the GraphQL layer as we have to manually input `company.credit_line` as its decimal equivalent in the response. It makes sense to change the repo layer if another presentation layer like a CLI needs to work with decimals. How often are Elixir projects (or most projects for that matter), using multiple presentation layers though? I'm ultimately also on the fence about this because the other generated tests do things like hardcode the use of 42/43 so we're not really deviating all that far from Phoenix norms.