#8 Company available_credit applied on transaction create or update
Add behavior to calculate the `available_credit` when a transaction is created though I just realized transaction updates require a diff check to adjust based on the difference between the current and new amounts. We need to do a similar check when updating a company's credit_line, allowing for negatives.
There's likely a more elegant way to apply a + or negative based on credit/debit boolean states but this is sufficient for now.