#8 Company available_credit applied to create and update
Adjust the behavior to use the `credit_line` on insert and apply the difference between `credit_line` values on update. If you adjust the credit_line + $100, the available_credit should also add $100 since you should always be working with the running total. I suspect there may be other dragons with this approach and it unfortunately takes the `(id, amount)` arity we had for transactions.

Transactions need to take a similar approach. On creation the amount should be adjusted. When a transaction is updated, the difference should be applied there as well.