#8 Company available_credit completed
Fix bug where I incorrectly closed the attributes map early. No clue how long that gave me grief.
Add `apply_available_credit_amount/2` function to handle the changes from transactions. We're always doing addition while letting the parent determine if it's a credit (add) or debit (subtraction).
Transaction functions refactored to apply the amount directly on creation or use the difference on update. Still feel like there's a better way to handle that if statement but at least all our tests pass.
Enabling the company tests exposed a flaw in my logic where on update we need to apply the difference to the `available_credit` and not the `credit_line`. We're essentially doing the math twice. Will look to correct this tomorrow and hopefully turn it in.