Write up complete
Mark objectives and subtasks as complete.
Finish time break down and total calculation.
Write up summaries of each objective, how I approached them and felt overall. I'm a huge fan of
this kind of assignment where boilerplate is provided. It gives me an idea of what to to expect
in a possible production code base.